Move all non-core sampling methods into the Distributions mixin.
